+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Homeworlds: Added `no-moves` descriptor to disable move generation on the front end.
 * Fix sameMove when move1 is a game ending move.
 * Martian Chess: Fixed a bug that occurred when someone resigned.
+* Fendo: Fixed edge case where a trapped piece couldn't build a wall.
 
 ### Changed